Create README.md
A mother of 6. I live in the United States.
Interested in writing articles about real life events, creating an investigating unit to help in the search for missing children and adults. I also carry an interest in testing websites and software. I am currently learning the True history of how America was established and I'm Looking to collaborate with creators who possess the same interest.